9주차 Day 2

+++ controller/CartController.js
@@ -1,10 +1,8 @@
 const jwt = require('jsonwebtoken');
 const conn = require('../mariadb') //db 모듈
 const {StatusCodes} = require('http-status-codes'); //status code 모듈
-const dotenv = require('dotenv');
 const ensureAuthorization = require('../auth'); //인증 모듈
 
-dotenv.config();
 
 //장바구니 담기
 const addCart = (req, res) => {
@@ -60,48 +58,52 @@ const getCartItems = (req, res) => {
             message: "잘못된 토큰입니다."
         });
     }
-    
-    if(selected)
-    {
-        let sql = `SELECT cart_items.id, book_id, title, summary, quantity, price
-            FROM cart_items LEFT JOIN books 
-            ON cart_items.book_id = books.id 
-            WHERE user_id = ? AND cart_items.id IN (?)`;
-        let values = [authorization.id, selected];
-        conn.query(sql, values,
-            function(err, results)
-            {
-                if(err)
-                {
-                    return res.status(StatusCodes.BAD_REQUEST).json(err); 
-                }
 
-                return res.status(StatusCodes.OK).json(results);
-            }
-        );
-    }
-    else {
-        let sql = `SELECT cart_items.id, book_id, title, summary, quantity, price
+    let sql = `SELECT cart_items.id, book_id, title, summary, quantity, price
             FROM cart_items LEFT JOIN books 
             ON cart_items.book_id = books.id 
             WHERE user_id = ?`;
-        conn.query(sql, authorization.id,
-            function(err, results)
-            {
-                if(err)
-                {
-                    return res.status(StatusCodes.BAD_REQUEST).json(err); 
-                }
-
-                return res.status(StatusCodes.OK).json(results);
-            }
-        );
+    values = [authorization.id];
+    
+    if(selected) //주문서 작성 시 '선택한 장바구니 목록 조회'
+    {
+        sql += ` AND cart_items.id IN (?)`;
+        values.push(selected);
     }
+
+    conn.query(sql, values,
+        function(err, results)
+        {
+            if(err)
+            {
+                return res.status(StatusCodes.BAD_REQUEST).json(err); 
+            }
+
+            return res.status(StatusCodes.OK).json(results);
+        }
+    );
 
 };
 
 //장바구니 도서 삭제
 const removeCartItem = (req, res) => {
+
+    let authorization = ensureAuthorization(req);
+
+    if(authorization instanceof jwt.TokenExpiredError)
+    {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            message: "로그인 세션이 만료되었습니다. 다시 로그인하세요."
+        });
+    }
+
+    if(authorization instanceof jwt.JsonWebTokenError)
+    {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            message: "잘못된 토큰입니다."
+        });
+    }
+
     const cartItemId = req.params.id;
 
     let sql = `DELETE FROM cart_items WHERE id = ?`;
